+++ secfsdstools/f_standardize/base_rule_framework.py
@@ -127,7 +127,7 @@ class AbstractRule(RuleEntity):
 class PrePivotRule(AbstractRule):
     """Base class to define a single rule that is applied before the dataframe was pivoted"""
 
-    index_cols = ['adsh', 'coreg', 'report', 'ddate', 'uom', 'qtrs', 'tag', 'version']
+    index_cols = ['adsh', 'coreg', 'report', 'ddate', 'qtrs', 'tag', 'version']
 
     def __init__(self, rule_id: str):
         self.rule_id = rule_id