+++ secfsdstools/f_standardize/standardizing.py
@@ -15,7 +15,6 @@ STANDARDIZED = TypeVar('STANDARDIZED', bound='StandardizedBag')
 
 LOGGER = logging.getLogger(__name__)
 
-
 class StandardizedBag:
     """
     A class to contain the results of a standardizer.
@@ -100,6 +99,71 @@ class StandardizedBag:
                                process_description_df=process_description_df)
 
 
+    @staticmethod
+    # pylint: disable=R0914
+    def concat(bags: List[STANDARDIZED]) -> STANDARDIZED:
+        """
+        Concat multiple StandardizedBags into a single one
+        Args:
+            bags: List of StandardizeBag instances to be concatenated
+
+        Returns:
+            StandardizedBag: single result
+
+        """
+
+        result_dfs = [bag.result_df for bag in bags]
+        applied_prepivot_rules_log_dfs = [bag.applied_prepivot_rules_log_df for bag in bags]
+        applied_rules_log_dfs = [bag.applied_rules_log_df for bag in bags]
+
+        # get stats_df, without the _rel and _gain cols
+        stats_dfs = [bag.stats_df.loc[:, ~bag.stats_df.columns.str.endswith('_rel') &
+                                         ~bag.stats_df.columns.str.endswith('_gain')]
+                     for bag in bags]
+
+        applied_rules_sum_ss = [bag.applied_rules_sum_s for bag in bags]
+
+        # get validation_overview_dfs without _pct column
+        validation_overview_dfs = [bag.validation_overview_df.loc[:,
+                                   ~bag.validation_overview_df.columns.str.endswith('_pct')]
+                                   for bag in bags]
+        process_description_dfs = [bag.process_description_df for bag in bags]
+
+        result_df = pd.concat(result_dfs, ignore_index=True)
+        applied_prepivot_rules_log_df = pd.concat(applied_prepivot_rules_log_dfs, ignore_index=True)
+        applied_rules_log_df = pd.concat(applied_rules_log_dfs, ignore_index=True)
+        applied_rules_sum_s: pd.Series = sum(applied_rules_sum_ss)
+        process_description_df = process_description_dfs[0]
+
+        # handling stats
+        #  stats_dfs only contains stats_df objects without _rel and _gain, so we can simply sum
+        stats_df: pd.DataFrame = sum(stats_dfs)
+        # next we use the Stats class do recalculate the _gain and _rel columns
+        stats = Stats([])
+        stats.stats = stats_df
+        stats.finalize_stats(len(result_df))
+
+        stats_df = stats.stats
+
+        # handling validation overview
+        validation_overview_df = validation_overview_dfs[0]
+        for entry_df in validation_overview_dfs[1:]:
+            validation_overview_df = validation_overview_df.add(entry_df, fill_value=0)
+
+        # calculate validation percentage columns
+        for col in validation_overview_df.columns:
+            validation_overview_df[f"{col}_pct"] = \
+                100 * (validation_overview_df[col] / len(result_df))
+
+        return StandardizedBag(result_df=result_df,
+                               applied_prepivot_rules_log_df=applied_prepivot_rules_log_df,
+                               applied_rules_log_df=applied_rules_log_df,
+                               stats_df=stats_df,
+                               applied_rules_sum_s=applied_rules_sum_s,
+                               validation_overview_df=validation_overview_df,
+                               process_description_df=process_description_df)
+
+
 class Stats:
     """
     Simple class to hold the process statics. This class contains
@@ -176,7 +240,7 @@ class Standardizer(Presenter[JoinedDataBag]):
     """
 
     # this tags identify single statements in the final standardized table
-    identifier_cols = ['adsh', 'coreg', 'report', 'ddate', 'uom', 'qtrs']
+    identifier_cols = ['adsh', 'coreg', 'report', 'ddate', 'qtrs']
 
     def __init__(self,
                  prepivot_rule_tree: RuleGroup,
@@ -293,6 +357,11 @@ class Standardizer(Presenter[JoinedDataBag]):
     def _preprocess(self, data_df: pd.DataFrame) -> pd.DataFrame:
         # sourcery skip: simplify-len-comparison, use-named-expression
         # only select rows with tags that are actually used by the defined rules
+
+        currency_uoms = [x for x in list(data_df.uom.unique()) if (len(x) == 3) and x.isupper()]
+        if len(currency_uoms) > 1:
+            raise ValueError("Multiple currencies are not supported. "
+                             "Please make sure that the uom column only contains one currency.")
 
         relevant_pivot_cols = \
             self.identifier_cols + ['tag', 'version', 'value', 'line', 'negating']
@@ -456,7 +525,8 @@ class Standardizer(Presenter[JoinedDataBag]):
         """
         standardized_df = self.process(databag.pre_num_df)
 
-        data_to_merge_df = databag.sub_df[['adsh', 'cik', 'form', 'fye', 'fy', 'fp', 'filed']].copy()
+        data_to_merge_df = \
+            databag.sub_df[['adsh', 'cik', 'form', 'fye', 'fy', 'fp', 'filed']].copy()
 
         # The name of a company can change during its liftime. However, we want to have the
         # same name for the same cik in all entries. Therefore, we first have to find the